refactor: meal finder tool
