+++ app/templates/base.html
@@ -26,7 +26,7 @@
     <body>
         <div>FE16 Tools:
             <a href="{{ url_for('index') }}">Home</a>
-            <a href="{{ url_for('meal_finder') }}">Meal Finder</a>
+            <a href="{{ url_for('tools.meal_finder') }}">Meal Finder</a>
             <a href="{{ url_for('tea_helper') }}">Tea Helper</a>
             <a href="{{ url_for('item_helper') }}">Item Helper</a>
             <a href="{{ url_for('seed_simulator') }}">Seed Simulator</a>
