+++ app/tools/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('tools', __name__)
+
+from app.tools import forms, routes
